rust: safe integer arithmetic
Tested to work only for widening ops. Need similar code for
subtraction and negative numbers etc. Once we have it working
we should write similar code for other languages that need it.

The following cases pass on rust:

```
    assert add1(127  - 1, 1) == 127
    assert add2(32767, 1) == 32768
    assert add3(2147483647, 1) == 2147483648
```

Related to: https://github.com/adsharma/py2many/issues/123

+++ tests/ext_expected/string_sum.rs
@@ -20,7 +20,7 @@ use pyo3::wrap_pyfunction;
 
 #[pyfunction]
 pub fn sum_as_string(a: i8, b: i8) -> PyResult<String> {
-    return Ok((a + b).to_string());
+    return Ok((a as i16 + b as i16).to_string());
 }
 
 #[pyfunction]
